debian autoinstall

+++ config/debian_install.sh
@@ -1,16 +1,16 @@
 #!/bin/sh
 
 apt-get install curl
-curl http://downloads.basho.com.s3-website-us-east-1.amazonaws.com/riak/CURRENT/debian/6/riak_1.2.1-1_amd64.deb -o riak_1.2.1-1_amd64.deb
-curl http://www.rabbitmq.com/releases/rabbitmq-server/v2.8.4/rabbitmq-server_2.8.4-1_all.deb -o rabbitmq-server_2.8.4-1_all.deb
-curl wget http://ftp.mirror.tw/pub/apache/lucene/solr/4.0.0/apache-solr-4.0.0.tgz -o apache-solr-4.0.0.tgz
+[ -f riak_1.2.1-1_amd64.deb ] || curl http://downloads.basho.com.s3-website-us-east-1.amazonaws.com/riak/CURRENT/ubuntu/precise/riak_1.2.1-1_amd64.deb -o riak_1.2.1-1_amd64.deb
+[ -f rabbitmq-server_2.8.4-1_all.deb ] || curl http://www.rabbitmq.com/releases/rabbitmq-server/v2.8.4/rabbitmq-server_2.8.4-1_all.deb -o rabbitmq-server_2.8.4-1_all.deb
+[ -f apache-solr-4.0.0.tgz ] || curl  http://apache.cdpa.nsysu.edu.tw/lucene/solr/4.0.0/apache-solr-4.0.0.tgz -o apache-solr-4.0.0.tgz
 tar xvfz apache-solr-4.0.0.tgz
-wget http://nodejs.org/dist/v0.8.4/node-v0.8.4.tar.gz
-tar xvfz node-v0.8.4.tar.gz
-cd node-v0.8.4
-./configure
-make
-make install
+#wget http://nodejs.org/dist/v0.8.4/node-v0.8.4.tar.gz
+#tar xvfz node-v0.8.4.tar.gz
+#cd node-v0.8.4
+#./configure
+#make
+#make install
 curl http://npmjs.org/install.sh | sh
 cd ..
 apt-get install build-essential openjdk-6-jdk  libssl-dev pkg-config curl libcurl4-openssl-dev php5-curl php5-pecl php5-memcache memcached supervisor git
@@ -25,6 +25,7 @@ apt-get install munin munin-node
 pecl install -n solr
 apt-get -f install
 dpkg -i riak_1..2.1-1_amd64.deb
+[ -f node-v0.6.18.tar.gz ] || wget http://nodejs.org/dist/v0.8.4/node-v0.6.18.tar.gz
 tar xvfz node-v0.6.18.tar.gz
 cd node-v0.6.18/
 ./configure && make && make install && supervisorctl restart all
